+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added HTTP Basic Authentication support to SemEHR API endpoint connection.
+
 
 ## [5.0.9] 2022-05-27